bringing firebase back online
