+++ components/lazy/schedules/index.tsx
@@ -2,7 +2,7 @@ import { useSessionValue } from 'hooks/session'
 import { lazy } from 'react'
 
 const Demo = lazy(() => import('./demo'))
-//const Firebase = lazy(() => import('./firebase'))
+const Firebase = lazy(() => import('./firebase'))
 const Localstorage = lazy(() => import('./localstorage'))
 
 export default function LazySchedules() {
@@ -10,8 +10,8 @@ export default function LazySchedules() {
   switch (session) {
     case 'demo':
       return <Demo />
-    //case 'firebase':
-    //return <Firebase />
+    case 'firebase':
+      return <Firebase />
     case 'localstorage':
       return <Localstorage />
     default:
