+++ hooks/schedules/demo.ts
@@ -2,7 +2,10 @@ import { schedules } from 'database/demo'
 import { nanoid } from 'nanoid'
 import create from 'zustand'
 import type {
-  AddSchedule, DeleteSchedule, EditSchedule, Schedules
+  AddSchedule,
+  DeleteSchedule,
+  EditSchedule,
+  Schedules,
 } from './types'
 import { addSchedule, deleteSchedule, editSchedule } from './utils'
 
@@ -18,27 +21,27 @@ const useStore = create<StoreState>((set) => ({
   addSchedule: async (schedule) => {
     const id = nanoid()
     set(({ schedules }) => ({
-      schedules: addSchedule(schedules, {...schedule, id}),
+      schedules: addSchedule(schedules, { ...schedule, id }),
     }))
     return { id }
   },
   editSchedule: async (schedule, id) =>
-  set(({ schedules }) => ({
-    schedules: editSchedule(schedules, schedule, id),
-  })),
+    set(({ schedules }) => ({
+      schedules: editSchedule(schedules, schedule, id),
+    })),
   deleteSchedule: async (id) =>
-  set(({ schedules }) => ({
-    schedules: deleteSchedule(schedules, id),
-  })),
+    set(({ schedules }) => ({
+      schedules: deleteSchedule(schedules, id),
+    })),
 }))
 
 const selectSchedules = (state: StoreState) => state.schedules
 export const useSchedules = () => useStore(selectSchedules)
 
-const selectAddSchedule=(state: StoreState) => state.addSchedule
+const selectAddSchedule = (state: StoreState) => state.addSchedule
 export const useAddSchedule = () => useStore(selectAddSchedule)
 
-const selectEditSchedule =(state: StoreState) => state.editSchedule
+const selectEditSchedule = (state: StoreState) => state.editSchedule
 export const useEditSchedule = () => useStore(selectEditSchedule)
 
 const selectDeleteSchedule = (state: StoreState) => state.deleteSchedule
