+++ package.json
@@ -2,6 +2,8 @@
   "scripts": {
     "build": "next build",
     "build-storybook": "STORYBOOK=true build-storybook",
+    "lint": "eslint . --ext ts,tsx,js,jsx",
+    "lint:strict": "npm run lint -- --max-warnings 0",
     "dev": "next",
     "storybook": "STORYBOOK=true start-storybook -p 6006",
     "start": "next start"
@@ -23,6 +25,18 @@
     "zustand": "^3.5.1"
   },
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^4.22.0",
+    "@typescript-eslint/parser": "^4.22.0",
+    "babel-eslint": "^10.0.0",
+    "eslint": "^7.24.0",
+    "eslint-config-react-app": "^6.0.0",
+    "eslint-plugin-flowtype": "^5.7.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jest": "^24.3.5",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-react": "^7.23.2",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-plugin-testing-library": "^4.1.0",
     "@storybook/addon-actions": "^6.2.9",
     "@storybook/addon-essentials": "^6.2.9",
     "@storybook/addon-links": "^6.2.9",
